refactored variables
