+++ lib/ipinfo-rails/ip_selector/ip_selector_interface.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+
 module IPSelectorInterface
     class InterfaceNotImplemented < StandardError; end
     def get_ip()
